refactor: Move flatten_pattern_elements() and create_locale_plural_variants() to transformer.py, add type hints
