+++ pontoon/pretranslation/tests/test_transformer.py
@@ -1,7 +1,11 @@
-from fluent.syntax import FluentParser
+import pytest
+from fluent.syntax import FluentParser, FluentSerializer
 from textwrap import dedent
 
-from pontoon.pretranslation.transformer import PretranslationTransformer
+from pontoon.pretranslation.transformer import (
+    PretranslationTransformer,
+    create_locale_plural_variants,
+)
 
 
 def visit(src):
@@ -126,3 +130,43 @@ def test_transformer_callback():
 
     assert len(ast.value.elements) == 1
     assert ast.value.elements[0].value == "res: My Title"
+
+
+@pytest.mark.django_db
+def test_create_locale_plural_variants(locale_a):
+    # Create default locale plural variants
+    input = dedent(
+        """
+        my-entry =
+            { $num ->
+                [0] Yo!
+                [one] Hello!
+               *[other] { reference } World!
+            }
+    """
+    )
+
+    parser = FluentParser()
+    message = parser.parse_entry(input)
+    expression = message.value.elements[0].expression
+
+    locale_a.cldr_plurals = "1,2,3,5"
+    create_locale_plural_variants(expression, locale_a)
+
+    expected = dedent(
+        """
+        my-entry =
+            { $num ->
+                [0] Yo!
+                [one] Hello!
+                [two] { reference } World!
+                [few] { reference } World!
+               *[other] { reference } World!
+            }
+    """
+    )
+
+    serializer = FluentSerializer()
+    assert serializer.serialize_entry(message) == serializer.serialize_entry(
+        parser.parse_entry(expected)
+    )
